requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.27.2
pydantic==2.11.3
pydantic-settings==2.8.1
python-dotenv==1.0.1
orjson==3.10.16
librosa==0.10.2.post1
soundfile==0.13.1
openai-whisper==20240930
ffmpeg-python==0.2.0
